refactored states to go on the top container and data to be passed as props
